Address easy comments on PR
