+++ README.md
@@ -506,47 +506,25 @@ Both ExpoArkProvider and ExpoIndexerProvider are available as adapters following
 - **ExpoArkProvider**: Handles settlement events and transaction streaming using expo/fetch for Server-Sent Events
 - **ExpoIndexerProvider**: Handles address subscriptions and VTXO updates using expo/fetch for JSON streaming
 
-To use IndexedDB repositories in Expo/React Native, install `indexeddbshim` and use
-the SDK's WebSQL-over-expo-sqlite adapter exported from `@arkade-os/sdk/adapters/expo`.
-Wire it in before importing the SDK so indexeddbshim can find `openDatabase`:
+To use IndexedDB repositories in Expo/React Native, call `setupExpoDb()` before any SDK import.
+This sets up `indexeddbshim` backed by expo-sqlite under the hood:
 
 ```typescript
-/**
- * IndexedDB polyfill for React Native (iOS / Android).
- *
- * Uses indexeddbshim backed by our WebSQL-over-expo-sqlite adapter.
- * Loaded via Metro's .native.ts platform extension — on web the
- * no-op sibling (indexeddb.ts) is loaded instead.
- *
- * Must be imported AFTER the crypto polyfill and BEFORE any SDK import.
- */
+import { setupExpoDb } from '@arkade-os/sdk/adapters/expo-db';
 
-import { openDatabase } from '@arkade-os/sdk/adapters/expo';
-
-// 1. Ensure `window` is defined — the SDK's getGlobalObject() and
-//    indexeddbshim both reference it.
-if (typeof (globalThis as any).window === 'undefined') {
-  (globalThis as any).window = globalThis;
-}
-
-// 2. Stub location.origin — indexeddbshim reads it for origin checks.
-if (typeof (globalThis as any).location === 'undefined') {
-  (globalThis as any).location = { origin: 'expo://localhost' };
-}
-
-// 3. Attach our WebSQL adapter so the shim finds openDatabase.
-(globalThis as any).openDatabase = openDatabase;
-
-// 4. Initialize indexeddbshim — sets globalThis.indexedDB + all IDB* globals.
-// @ts-ignore — no types for the browser-noninvasive entry
-import setGlobalVars from 'indexeddbshim';
-
-setGlobalVars(globalThis, {
-  checkOrigin: false,
-  useSQLiteIndexes: true,
-  cacheDatabaseInstances: true,
-});
+setupExpoDb();
 ```
+
+> **Note:** `setupExpoDb` accepts an optional `SetupExpoDbOptions` object to
+> customise `origin`, `checkOrigin`, and `cacheDatabaseInstances`.
+
+> **Note:** `expo-sqlite` and `indexeddbshim` are optional peer dependencies,
+> only required when importing from `@arkade-os/sdk/adapters/expo-db`. The
+> streaming providers (`@arkade-os/sdk/adapters/expo`) have no expo-sqlite
+> dependency. Install them with:
+> ```bash
+> npx expo install expo-sqlite && npm install indexeddbshim
+> ```
 
 #### Crypto Polyfill Requirement
